chroma db cleaned and organized
