+++ api/urls.py
@@ -5,11 +5,11 @@ from .views.conversation_view import (
     PromptConversationAdminView,
 )
 
-from .views.brain_view import (
-    ListReviewAndUpdateBrainView,
-    UpdateReviewStatusView,
-    UpdateBrainView,
-)
+# from .views.brain_view import (
+    # ListReviewAndUpdateBrainView,
+    # UpdateReviewStatusView,
+    # UpdateBrainView,
+# )
 
 # Define URL patterns
 urlpatterns = [
@@ -17,7 +17,7 @@ urlpatterns = [
     path("prompt_conversation/", PromptConversationView.as_view(), name="prompt_conversation"),
     path("prompt_conversation_admin/", PromptConversationAdminView.as_view(), name="prompt_conversation_admin",),
     # Brain
-    path("list_review_update_brain/", ListReviewAndUpdateBrainView.as_view(), name="review_update_brain"),
-    path("update_review_status/", UpdateReviewStatusView.as_view(), name="update_review_status"),
-    path("update_brain/", UpdateBrainView.as_view(), name="update_brain"),
+    # path("list_review_update_brain/", ListReviewAndUpdateBrainView.as_view(), name="review_update_brain"),
+    # path("update_review_status/", UpdateReviewStatusView.as_view(), name="update_review_status"),
+    # path("update_brain/", UpdateBrainView.as_view(), name="update_brain"),
 ]
